Move mallocspace to marksweepspace as a module.
