--- src/policy/mallocspace/mod.rs
@@ -1,5 +1,0 @@
-///! A marksweep space that allocates from malloc.
-mod global;
-pub mod metadata;
-
-pub use global::*;
