+++ src/policy/mod.rs
@@ -25,6 +25,6 @@ pub mod immix;
 pub mod immortalspace;
 pub mod largeobjectspace;
 pub mod lockfreeimmortalspace;
-pub mod mallocspace;
+// pub mod mallocspace;
 pub mod markcompactspace;
 pub mod marksweepspace;
