+++ src/util/alloc/mod.rs
@@ -39,4 +39,4 @@ pub use markcompact_allocator::MarkCompactAllocator;
 /// Embedded metadata pages
 pub(crate) mod embedded_meta_data;
 
-pub use crate::policy::mallocspace::metadata::is_alloced_by_malloc;
+pub use crate::policy::marksweepspace::malloc_ms::is_alloced_by_malloc;
